Code review comment fixes
